requirements.txt
python-binance==1.0.16
pydantic==1.10.11
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

